+++ junit/kube-api-test/core/src/test/java/io/fabric8/kubeapitest/sample/KubernetesMutationHookHandlingTest.java
@@ -16,7 +16,6 @@
 package io.fabric8.kubeapitest.sample;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import io.fabric8.kubeapitest.cert.CertManager;
 import io.fabric8.kubeapitest.junit.EnableKubeAPIServer;
 import io.fabric8.kubeapitest.junit.KubeConfig;
 import io.fabric8.kubernetes.api.model.HasMetadata;
@@ -42,12 +41,10 @@ import io.fabric8.mockwebserver.http.Dispatcher;
 import io.fabric8.mockwebserver.http.MockResponse;
 import io.fabric8.mockwebserver.http.RecordedRequest;
 import io.fabric8.zjsonpatch.JsonDiff;
-import org.bouncycastle.asn1.x509.GeneralName;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
@@ -69,22 +66,10 @@ class KubernetesMutationHookHandlingTest {
   static String kubeConfig;
 
   static MockWebServer webhookServer;
-  static File certFile;
 
   @BeforeAll
   static void setupWebhookServer() {
-    certFile = new File("target/conversion.crt");
-    if (certFile.getParentFile() != null) {
-      certFile.getParentFile().mkdirs();
-    }
-
-    File keyFile = new File("target/conversion.key");
-
-    CertManager.generateKeyAndCertificate("CN=localhost", keyFile, certFile,
-        new GeneralName(GeneralName.dNSName, "localhost"));
-
     webhookServer = new MockWebServer();
-    webhookServer.usePemCertificates(keyFile, certFile);
     webhookServer.useHttps();
     webhookServer.setDispatcher(new Dispatcher() {
       @Override
@@ -96,7 +81,7 @@ class KubernetesMutationHookHandlingTest {
         return new MockResponse().setResponseCode(404);
       }
     });
-    webhookServer.start(8443);
+    webhookServer.start(0);
   }
 
   @AfterAll
@@ -147,7 +132,9 @@ class KubernetesMutationHookHandlingTest {
     try (InputStream resource = KubernetesMutationHookHandlingTest.class
         .getResourceAsStream("/MutatingWebhookConfig.yaml")) {
       MutatingWebhookConfiguration hook = (MutatingWebhookConfiguration) client.load(resource).items().get(0);
-      hook.getWebhooks().get(0).getClientConfig().setCaBundle(WebhookServerTestUtils.getEncodedCertificate(certFile));
+      hook.getWebhooks().get(0).getClientConfig().setUrl("https://localhost:" + webhookServer.getPort() + "/mutate");
+      hook.getWebhooks().get(0).getClientConfig().setCaBundle(
+          WebhookServerTestUtils.getEncodedCertificate(webhookServer.getSelfSignedCertificate().certificatePath()));
       client.resource(hook).serverSideApply();
     } catch (IOException e) {
       throw new RuntimeException(e);